Move container source from provider to repository
