+++ config/services.yaml
@@ -9,42 +9,33 @@ parameters:
     google_cloud_credentials: '%env(json:GOOGLE_APPLICATION_CREDENTIALS)%'
 
 services:
-    # default configuration for services in *this* file
     _defaults:
         autowire: true      # Automatically injects dependencies in your services.
         autoconfigure: true # Automatically registers your services as commands, event subscribers, etc.
         public: false       # Allows optimizing the container by removing unused services; this also means
-        # fetching services directly from the container via $container->get() won't work.
-        # The best practice is to be explicit about your dependencies anyway.
 
-    # makes classes in src/ available to be used as services
-    # this creates a service per class whose id is the fully-qualified class name
     App\:
         resource: '../src/*'
         exclude: '../src/{DependencyInjection,Entity,Migrations,Tests,Kernel.php}'
 
-    # controllers are imported separately to make sure services can be injected
-    # as action arguments even if you don't extend any base controller class
     App\Controller\:
         resource: '../src/Controller'
         tags: ['controller.service_arguments']
+
+    App\Controller\Exception:
+        public: true
+        arguments:
+            $debug: '%kernel.debug%'
 
     App\Serializer\:
         resource: '../src/Serializer'
         tags: ['serializer.normalizer']
 
-    # add more service definitions when explicit configuration is needed
-    # please note that last definitions always *replace* previous ones
-
-    App\Provider\Github:
+    App\Service\Github:
         arguments:
             $githubUser: '%env(CONFIG_GITHUB_USER)%'
             $githubRepo: '%env(CONFIG_GITHUB_REPO)%'
             $githubBranch: '%env(CONFIG_GITHUB_BRANCH)%'
-
-    App\Provider\ContainerProvider:
-        arguments:
-            $path: '%containers_path%'
 
     ZipArchive:
         class: \ZipArchive
@@ -55,8 +46,3 @@ services:
         arguments:
             $clientId: '%env(GITHUB_APP_ID)%'
             $clientSecret: '%env(GITHUB_APP_SECRET)%'
-
-    App\Controller\Exception:
-        public: true
-        arguments:
-            $debug: '%kernel.debug%'
