+++ src/Repository/ContainerRepositoryInterface.php
@@ -6,6 +6,7 @@ namespace App\Repository;
 
 use App\Entity\Container;
 use App\ValueObject\Container\Manifest;
+use App\ValueObject\File;
 
 /**
  * @author Laurent Bassin <laurent@bassin.info>
@@ -17,4 +18,6 @@ interface ContainerRepositoryInterface
     public function findOneById(string $id): Container;
 
     public function findManifestByContainerId(string $id): Manifest;
+
+    public function findSource(string $containerId, string $filename): File;
 }
